Relative synergy movements and close until contact
Added the ability to close the hand (by moving one joint of every finger) until the fingertips come into contact with something. Also changed the synergy movements from absolute movements relative to a zero position of the joints to a relative movement with respect to the current joint values. This makes the movements smaller, but maybe reduces the contacts the fingers have to each other during movement on the real robot. Both parts are yet untested and some parts of the second change are commented out.