Structure refactoring
Now the general control structure is the following. The handover
controller sends service request to the different parts representing all
necessary steps for the pipeline. Once one request is finished, the next
is send. This enables an easier change to more sophisticated parts later
and should allow the execution of a stub pipeline alread. Nothing is
tested yet.