Prototype to test if grasp is stable
This prototype computes a cartesian path along the z-axis for the shadow
hand and compares the wrench forces along the z-axis of the force/torque
sensor in the left arm before and after. This is the simplest approach
and still needs to be tested. Also, the wiat for messages for the wrench
messages should be eliminated.