Mimic collector and env
WIP. Added ability to collect data curing ccm for the mimic env. The
mimic env tries to learn to close the hand similar to the hardcoded ccm
using synergies. The learning is done with a different reward than the
later one used, but it can be useful to test the general framework.