Added new PCA based model learning.
Changed the env to a version, where PCA values were learned by the
network and were used for the state as well. Further information about
this is in the READ.ME, along with instructions to build the workspace
new. train_ppo, model_mover and syn_grasp_gen were updated accordingly.
One model, that works very well, was already trained with the updated
MimicEnv. For it to work, however, the 4 joints needed to be limited on
each finger except the thumb.