Plan fixed hand movement with rh_grasp link
To be able to have bio_ik plan the requested goals relative to the link rh_grasp, which is a custom link not in the basic robot model, a new robot model with the new frame needs to be setup and loaded by bio_ik. This now allows plannign with this frame.