Renamed init_gripper_stub to initial_setup
No longer is a stub and is now responsible for setting up both arms and head instead of just the left arm. Added the ability to reset the right arm into a default configuration and also switch between two different fixed configurations for the left arm, based on the provided object_type.