Full.launch included in train. No auto launch
Commented out the autolaunch of the full pipeline when running full.launch. Now this needs to be manually called with a rosrun of launch_handover. Removed args in train according to previous commit. Now train includes full on launch, as the handover pipeline is required for future changes in env.