Attempt to improve error handling.
Rarely an error occuurs during training, mainly with the book, that reads "Failed to fetch current robot state". If seems to result from a call to fingers.get_current_joint_values(). This might be the cause, because finger joint states are not published at every iteration of joint_states. Some attempts have been made to mitigate that (see newly added comment in code), as well as an exception handling that should help for diagnostics, if the error still occurs.