Changed camera link
Now uses azure_kinect_argb_camera_link_urdf instead of azure_kinect_argb_camera_link. These are essentailly the same link, but the first one is written in the urdf, while the second is published by the camera itself. This runs into issues when running the pipeline in demo mode, when the camera doesn't exist.