Bugfixes, renames, Markers, Fixed poses
Initialising the stat monitor correctly, so the RobotCommander actually
monitors the correct state. Renamed the debug publishers to be more
consistent. Added Markers for the checked gripper and hand poses and
colored them realtive to the FK pos diff. Fixed the gripper and hand
poses used for the initial sampling instead of using the current pose of
the gripper, enabling the gripper to be in a different pose.