Shifted several parameters to the config file
Handover type and grasp mode are no longer arguments for the launch file or the launch_handover file. Instead, they are set through the config file. Now only the demo mode for pc is an argument for the launch file and the object type is the only argument for the launch_handover file. Additionally, modes like debug, analyse, etc. are no getting set through the config file as well, so no more unnecessary parameters for the init functions in the corresponding files or moving through every file to change the debug options.