Fixed workspace analysis error
When setting the value of the result for the first sampled transform for
every handover point, at a failure the value was 2 instead of 1,
resulting in the possibility of increasing the number of wrong solutions
by 1. Since the workspace analysis was made before this correction, it
has to be manuall adjusted during visualization, which is possible with
the recorded data.