Repaired cropbox filter
The cropbox filter was previously wrongly set. The orientation was set to default, which means it had the same orientation as the azure_kinect_rgb_camera_link frame. Now, it is properly rotated and translated to the l_gripper_tool_frame and the box bounds are set relative to that frame. Also some cleanup for the pc_crop function.